+++ tests/testthat/test-pull.r
@@ -1,5 +1,31 @@
 context("test-pull")
 
+
+test_that("pull with",  {
+  flights_df = as.disk.frame(nycflights13::flights)
+  
+  a = flights_df %>% 
+    pull(carrier, carrier)
+  
+  b = flights_df %>% collect() %>% pull(carrier, carrier)
+  
+  expect_equal(a, b)
+  
+  a = flights_df %>% 
+    pull(2, 2)
+  b = flights_df %>% collect() %>% pull(2, 2)
+  
+  expect_equal(a, b)
+  
+  a = flights_df %>% 
+    pull(-1, -1)
+  b = flights_df %>% collect() %>% pull(-1, -1)
+  expect_equal(a, b)
+  
+  delete(flights_df)
+})
+
+
 test_that("pull",  {
   flights_df = as.disk.frame(nycflights13::flights)